+++ trino-lb-core/src/trino_query.rs
@@ -52,6 +52,10 @@ pub struct TrinoQuery {
     /// Endpoint of the Trino cluster the query is running on.
     pub trino_endpoint: Url,
 
+    /// (Optionally, if configured) public endpoint of the Trino cluster.
+    /// This can e.g. be used to change segment ackUris to.
+    pub trino_external_endpoint: Option<Url>,
+
     /// The time the query was submitted to trino-lb.
     pub creation_time: SystemTime,
 
@@ -80,6 +84,7 @@ impl TrinoQuery {
         trino_cluster: TrinoClusterName,
         trino_query_id: TrinoQueryId,
         trino_endpoint: Url,
+        trino_external_endpoint: Option<Url>,
         creation_time: SystemTime,
         delivered_time: SystemTime,
     ) -> Self {
@@ -87,6 +92,7 @@ impl TrinoQuery {
             id: trino_query_id,
             trino_cluster,
             trino_endpoint,
+            trino_external_endpoint,
             creation_time,
             delivered_time,
         }